Fix unit tests

+++ src/gobapi/worker/api.py
@@ -1,9 +1,18 @@
+from time import sleep
 from flask import send_file, jsonify
 
 from gobapi.worker.response import WorkerResponse
 
 
 def worker_result(worker_id):
+    """
+    Get the result of a worker.
+
+    If the worker has finished the worker file is returned.
+
+    :param worker_id:
+    :return:
+    """
     filename = WorkerResponse.getResponseFile(worker_id)
     if filename:
         return send_file(filename)
@@ -14,11 +23,27 @@ def worker_result(worker_id):
 
 
 def worker_end(worker_id):
-    WorkerResponse.kill(worker_id)
-    return "", 204  # No Content
+    """
+    End a running worker
+
+    :param worker_id:
+    :return:
+    """
+    status = WorkerResponse.getStatus(worker_id)
+    if status:
+        WorkerResponse.kill(worker_id)
+        return "", 204  # No Content
+    else:
+        return _worker_not_found(worker_id)
 
 
 def worker_status(worker_id):
+    """
+    Returns the status of a worker: dict(id, status, size)
+
+    :param worker_id:
+    :return:
+    """
     status = WorkerResponse.getStatus(worker_id)
     if status:
         return jsonify(status)
@@ -27,4 +52,11 @@ def worker_status(worker_id):
 
 
 def _worker_not_found(worker_id):
+    """
+    Returns a generic worker not found Response.
+
+    :param worker_id:
+    :return:
+    """
+    sleep(1)  # penalty time to protect against brute force attacks
     return f"Worker {worker_id} not found", 404  # Not found
